requirements.txt
pandas>=1.5.0
tqdm>=4.64.0
PyYAML>=6.0
python-dateutil>=2.8.0
openpyxl>=3.0.0
xlrd>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
python-dotenv>=1.0.0